Extract command handlers into src/commands/ module
Move subcommand handlers (password, config, status, reset, verify,
import, list, login) and the 421 recovery/service initialization
cluster into a dedicated commands module. Pure code move with
visibility adjustments.

main.rs: 2888 -> 1457 lines. New module: 10 files under commands/.
