+++ src/commands/status.rs
@@ -1,0 +1,63 @@
+use crate::cli;
+use crate::config;
+use crate::state;
+use crate::state::StateDb;
+
+/// Run the status command.
+pub(crate) async fn run_status(
+    args: cli::StatusArgs,
+    globals: &config::GlobalArgs,
+    toml: Option<&config::TomlConfig>,
+) -> anyhow::Result<()> {
+    let db_path = super::super::get_db_path(globals, toml)?;
+
+    if !db_path.exists() {
+        println!("No state database found at {}", db_path.display());
+        println!("Run a sync first to create the database.");
+        return Ok(());
+    }
+
+    let db = state::SqliteStateDb::open(&db_path).await?;
+    let summary = db.get_summary().await?;
+
+    println!("State Database: {}", db_path.display());
+    println!();
+    println!("Assets:");
+    println!("  Total:      {}", summary.total_assets);
+    println!("  Downloaded: {}", summary.downloaded);
+    println!("  Pending:    {}", summary.pending);
+    println!("  Failed:     {}", summary.failed);
+    println!();
+
+    if let Some(started) = &summary.last_sync_started {
+        println!(
+            "Last sync started:   {}",
+            started.format("%Y-%m-%d %H:%M:%S UTC")
+        );
+    }
+    if let Some(completed) = &summary.last_sync_completed {
+        println!(
+            "Last sync completed: {}",
+            completed.format("%Y-%m-%d %H:%M:%S UTC")
+        );
+    }
+
+    if args.failed && summary.failed > 0 {
+        println!();
+        println!("Failed assets:");
+        let failed = db.get_failed().await?;
+        for asset in failed {
+            let last_seen = asset.last_seen_at.format("%Y-%m-%d %H:%M:%S");
+            println!(
+                "  {} ({}) - {} (attempts: {}, last seen: {})",
+                asset.filename,
+                asset.id,
+                asset.last_error.as_deref().unwrap_or("unknown error"),
+                asset.download_attempts,
+                last_seen
+            );
+        }
+    }
+
+    Ok(())
+}
